+++ code/win32/win_syscon.cpp
@@ -1222,13 +1222,13 @@ void Conbuf_AppendText( const char *msg )
 	bufLen = b - buffer;
 
 	// not enough space in buffer -> flush
-	if ( bufLen + conBufPos >= sizeof( conBuffer )-1 ) {
+	if ( bufLen + conBufPos >= static_cast<int>( sizeof( conBuffer ) ) - 1 ) {
 		AddBufferText( conBuffer, conBufPos );
 		conBufPos = 0;
 	}
 
 	// new message is too long -> flush
-	if ( bufLen >= sizeof( conBuffer )-1 ) {
+	if ( bufLen >= static_cast<int>( sizeof( conBuffer ) ) - 1 ) {
 		if ( conBufPos ) {
 			AddBufferText( conBuffer, conBufPos );
 			conBufPos = 0;
